Write good unit AND functional tests for contracts/Governance/Voting.sol #224
